requirements.txt
python-telegram-bot==20.6
SQLAlchemy==2.0.23
alembic==1.12.1
requests==2.31.0
qrcode==7.4.2
pillow
python-dotenv==1.0.0
cryptography==41.0.5
aiohttp==3.9.1
asyncio==3.4.3
psycopg2-binary
pytz==2023.3.post1
pydantic
uvicorn==0.24.0.post1
fastapi==0.104.1
python-multipart==0.0.6
bcrypt==4.0.1
redis==5.0.1
aioredis==2.0.1
schedule==1.2.1
APScheduler==3.10.4 